+++ docs/CHANGELOG.md
@@ -22,6 +22,7 @@
 - frontend: jump to last shown element when back to list 
 - frontend: set perPage=10 elements on search-link in inlinesearch
 - frontend: added backToTop-link
+- frontend: added parameters for configuring show-layout via url-parameter 
 - frontend: improved code
 - backend: added more facets/filters/fields to mediadb-adapter
 - backend: improved sdoc-loader/exporter